+++ Core/OfficeDevPnP.Core/Framework/Provisioning/ObjectHandlers/ObjectHierarchyTenant.cs
@@ -65,7 +65,10 @@ namespace OfficeDevPnP.Core.Framework.Provisioning.ObjectHandlers
                                 (hierarchy.Tenant.SiteScripts != null && hierarchy.Tenant.SiteScripts.Count > 0) ||
                                 (hierarchy.Tenant.StorageEntities != null && hierarchy.Tenant.StorageEntities.Count > 0) ||
                                 (hierarchy.Tenant.WebApiPermissions != null && hierarchy.Tenant.WebApiPermissions.Count > 0) ||
-                                (hierarchy.Tenant.Themes != null && hierarchy.Tenant.Themes.Count > 0)
+                                (hierarchy.Tenant.Themes != null && hierarchy.Tenant.Themes.Count > 0) ||
+                                (hierarchy.Tenant.SPUsersProfiles != null && hierarchy.Tenant.SPUsersProfiles.Count > 0) ||
+                                (hierarchy.Tenant.Office365GroupLifecyclePolicies != null && hierarchy.Tenant.Office365GroupLifecyclePolicies.Count > 0) ||
+                                (hierarchy.Tenant.Office365GroupsSettings?.Properties != null && hierarchy.Tenant.Office365GroupsSettings?.Properties.Count > 0)
                                 );
             }
             return (_willProvision.Value);